Fixed code

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@ int print_last_digit(int n)
 {
 	if (n < 0)
 	{
-		n = n * (-1);
+		n = (long) n * (-1);
 	}
 	_putchar(48 + (n % 10));
